feat: add skip same path
