+++ src/add-configuration.ts
@@ -1,6 +1,7 @@
 import linker from "./api/linker";
 import query from "./api/query";
-import { type ConfigParams, getSettings } from "./settings";
+import { getSettings } from "./settings";
+import type { ConfigParams } from "./types/settings";
 
 function mergeDefaults(config: ConfigParams = {}): ConfigParams {
   return {
