+++ src/add-router-tracking.ts
@@ -14,7 +14,11 @@ export default async function addRouterTracking(): Promise<void> {
 
   trackRoute(router.currentRoute.value);
 
-  router.afterEach((to) => {
+  router.afterEach((to, from) => {
+    if (to.path === from.path && pageTracker.skipSamePath) {
+      return;
+    }
+
     trackRoute(to);
   });
 }
